Rename local-storage test helper to web-storage
