+++ addon-test-support/-private/web-storage.ts
